requirements.txt
python-dotenv==1.0.0
wikipedia==1.4.0
tiktoken==0.5.1
neo4j==5.2.1
streamlit==1.27.2
sentence_transformers==2.2.2
transformers==4.34.1
Pillow==10.1.0
fastapi==0.103.2
PyPDF2==3.0.1
torch==2.0.1
pydantic==1.10.13
uvicorn==0.23.2
sse-starlette==1.6.5
numexpr==2.8.7
langchain==0.0.323
